library/backend/requirements.txt
# SmartLib Kiosk - Core System
# Face Recognition + Book Detection + Transaction Management

# Core Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
# Python 3.13 needs newer pydantic/pydantic-core wheels (avoid Rust build)
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart==0.0.6
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0
# Async Postgres driver for SQLAlchemy
asyncpg>=0.30.0
# Need a cp313 wheel on Windows (avoid building from source)
greenlet>=3.1.0
pgvector>=0.2.0

# Env loading (used by helper scripts)
python-dotenv>=1.0.0

# AI/ML - Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Face Recognition (Deep Neural Networks)
insightface==0.7.3
# onnxruntime-gpu>=1.16.0
onnxruntime>=1.16.0  # Use this for CPU-only

# Object Detection (YOLOv8)
ultralytics>=8.0.0

# OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Barcode
pyzbar>=0.1.9

# Image Processing
scikit-image>=0.21.0
albumentations>=1.3.0
faiss-cpu>=1.8.0

# API & Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Logging
loguru>=0.7.0
